Rewrites new upload files
