+++ apps/judicial-system/web/src/routes/Prosecutor/components/CaseFiles/CaseFiles.tsx
@@ -33,6 +33,7 @@ import {
 import PageHeader from '@island.is/judicial-system-web/src/components/PageHeader/PageHeader'
 import { CaseOrigin } from '@island.is/judicial-system-web/src/graphql/schema'
 import {
+  addUploadFiles,
   generateSingleFileUpdate,
   mapCaseFileToUploadFile,
   removeTabsValidateAndSet,
@@ -264,7 +265,10 @@ export const CaseFiles: React.FC<React.PropsWithChildren<unknown>> = () => {
               header={formatMessage(strings.filesLabel)}
               buttonLabel={formatMessage(strings.filesButtonLabel)}
               onChange={(files) =>
-                handleUpload(files, setDisplayFiles, handleUIUpdate)
+                handleUpload(
+                  addUploadFiles(files, setDisplayFiles),
+                  handleUIUpdate,
+                )
               }
               onRemove={(file) => handleRemove(file, removeFileCB)}
               onRetry={(file) => handleRetry(file, handleUIUpdate)}
