+++ apps/judicial-system/web/src/routes/Shared/Statement/Statement.tsx
@@ -30,6 +30,7 @@ import {
 import RulingDateLabel from '@island.is/judicial-system-web/src/components/RulingDateLabel/RulingDateLabel'
 import { CaseAppealDecision } from '@island.is/judicial-system-web/src/graphql/schema'
 import {
+  addUploadFiles,
   generateSingleFileUpdate,
   mapCaseFileToUploadFile,
 } from '@island.is/judicial-system-web/src/utils/formHelper'
@@ -157,16 +158,12 @@ const Statement = () => {
                 buttonLabel={formatMessage(core.uploadBoxButtonLabel)}
                 onChange={(files) =>
                   handleUpload(
-                    files,
-                    setDisplayFiles,
+                    addUploadFiles(files, setDisplayFiles, appealStatementType),
                     handleUIUpdate,
-                    appealStatementType,
                   )
                 }
                 onRemove={(file) => handleRemove(file, removeFileCB)}
-                onRetry={(file) =>
-                  handleRetry(file, handleUIUpdate, appealStatementType)
-                }
+                onRetry={(file) => handleRetry(file, handleUIUpdate)}
               />
             </Box>
             <Box component="section" marginBottom={10}>
@@ -189,16 +186,12 @@ const Statement = () => {
                 buttonLabel={formatMessage(core.uploadBoxButtonLabel)}
                 onChange={(files) =>
                   handleUpload(
-                    files,
-                    setDisplayFiles,
+                    addUploadFiles(files, setDisplayFiles, appealCaseFilesType),
                     handleUIUpdate,
-                    appealCaseFilesType,
                   )
                 }
                 onRemove={(file) => handleRemove(file, removeFileCB)}
-                onRetry={(file) =>
-                  handleRetry(file, handleUIUpdate, appealCaseFilesType)
-                }
+                onRetry={(file) => handleRetry(file, handleUIUpdate)}
               />
             </Box>
           </>
